Updates to Licensee Application new and thank you
This update fixes the hero image height (as well as add a few hero image
versions).

Additionally, the thank you redirect was not working due to a missing
action and route. These are added as well as some basic copy for the
thank you page.

Last, a small layout change is added to the new page so that the copy
'Apply Now' sits a little higher on the page.