Update Private Events new form
This commit begins the updates to the private_events#new page. This is a
work in progress.

In this first commit sees the addition of Bootstrap Datepicker,
Bootstrap Clockpocker, and Font Awesome. The form is reworked to fit the
new layout and date and time elements are updated to use the new JS
libraries.