Update Footer links
This commit moves the Partner with Us link to the site footer as well as
add links to the Terms and Privacy pages.

Basic files are also added for the terms and privacy pages. Currently
the copy for these pages are in a docx format and not easy to copy over
with formatting.

I am holding off on additional work on these pages to see if I can get
the copy in plain text format.