Add Terms and Privacy Policy copy
This update brings in the first pass at the terms and privacy_policy
pages copy layout.

An additional change of note is the removal of z-index on the footer
background image. Not sure why it was originally there as I cannot
locate an issues with its removal but leaving it present drops the image
behind the rest of the canvas on the privacy_policy page.