Add phone number and update page elements
This commit adds the provided phone number to the Contact page and site
footer.

Additionally, this commit addresses the gray header h1 text that no
longer fits the site design. The header h1's are hidden via bootstrap
helper class (which maintains screen reader visibility) and a new styled
h1 is added to the page copy.