Update Experiences index page
This commit makes several design changes to the Experiences index for
better aesthetics and usability.

The Experiences index page was also subject the gray h1 text, which is
removed. The page title is relocated outside the header and styled with
the site purple color for consistency.

This page never had a proper header image which resulted in a white on
white nav. The header image from the FAQ page is added to facilitate
proper display.

Lastly, the links within the class images in the default view are
reworked and expanded to include the full image, not just the title.
This makes clicking on classes much easier on the user.