Update orders#create action and orders#show page
Removing the `@price` instance variable as the refactoring to
`CheckPromo` consumed the mutation to tally the final price. In its
place, `total` is now used.

Additionally, the `orders#show` page did not have a header image which
resulted in the nav being hard to read (white on light gray). Adding in
the image from the contact page to allow the nav a better display and
conform to the current practice of always having a header image.
`.titleize` also replaces the `.capitalize` method for proper
capitalization of names.

And lastly, adding a blank js file `active_admin.js` to eliminate the
log error due Active Admin expecting cannot its presence.