Add Rollbar Gem for error tracking
While the addition is to gain insight into recent email issues, having
some form of error tracking in a production application is of great use
for general uptime and support.

Currently Rollbar is tied to D.Moore's account but at some point a
Merlot's Art account should be created and access provided to the
development team. This will allow up to 5000 events per month for free,
which should be plenty given the application's traffic.

With the present configuration, the 5000 event limit is shared across
multiple applications monitored by D.Moore.