Cast @public_events to Array to prevent nil
Errors were occurring on production when experience filters are combined
in a way that results in no `public_events` being returned. To correct
the problem quickly, the `@public_events` instance variable is cast to
an Array before calling `.each`. This will ensure the receiver will
always respond.

This is a band-aid fix. The filters are broken and need attention at
some point in the future.